Switched over to IntelliJ, added some initial UI code for download manager fragment

+++ SubsonicClient/src/com/casamento/subsonicclient/DownloadTask.java
@@ -1,28 +1,19 @@
 package com.casamento.subsonicclient;
 
-import java.io.BufferedInputStream;
-import java.io.File;
-import java.io.FileOutputStream;
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.ObjectInputStream;
-import java.io.OutputStream;
-import java.net.URISyntaxException;
-import java.net.URL;
-import java.net.URLConnection;
-
+import android.app.ProgressDialog;
+import android.os.AsyncTask;
+import android.util.Log;
 import org.apache.http.HttpEntity;
 import org.apache.http.HttpResponse;
-import org.apache.http.client.ClientProtocolException;
 import org.apache.http.client.HttpClient;
 import org.apache.http.client.methods.HttpGet;
 import org.apache.http.impl.client.DefaultHttpClient;
 
-import android.app.ProgressDialog;
-import android.os.AsyncTask;
-import android.util.Log;
+import java.io.*;
+import java.net.URISyntaxException;
+import java.net.URL;
 
-class DownloadTask extends AsyncTask<Void, Integer, String> {
+class DownloadTask extends AsyncTask<Void, Long, String> {
 	private final String logTag = "DownloadTask";
 	private URL url;
 	private String saveLocation;
@@ -45,10 +36,10 @@ class DownloadTask extends AsyncTask<Void, Integer, String> {
 	}
 	
 	@Override
-	protected void onProgressUpdate(Integer... progress) {
+	protected void onProgressUpdate(Long... progress) {
 		super.onProgressUpdate(progress);
 		
-		progressDialog.setProgress(progress[0]);
+		progressDialog.setProgress(progress[0].intValue());
 	}
 	
 	@Override
@@ -83,7 +74,7 @@ class DownloadTask extends AsyncTask<Void, Integer, String> {
 				while ((count = input.read(data)) != -1) {
 					total += count;
 					if (!this.indeterminate)
-						publishProgress((int)total);
+						publishProgress(total);
 					output.write(data, 0, count);
 				}
 				output.flush();
